# Reduce System.Mutate runtime by 87%