--- a/pkg/mutation/system.go
+++ b/pkg/mutation/system.go
@@ -105,7 +105,7 @@ func (s *System) Mutate(obj *unstructured.Unstructured, ns *corev1.Namespace) (b
 
 			if m.Matches(obj, ns) {
 				mutated, err := m.Mutate(obj)
-				if mutated && (*MutationLoggingEnabled || *MutationAnnotationsEnabled) {
+				if mutated {
 					appliedMutations = append(appliedMutations, m)
 				}
 				if err != nil {
@@ -113,6 +113,13 @@ func (s *System) Mutate(obj *unstructured.Unstructured, ns *corev1.Namespace) (b
 				}
 			}
 		}
+
+		if len(appliedMutations) == 0 {
+			// If no mutations were applied, we can safely assume the object is
+			// identical to before.
+			return i > 0, nil
+		}
+
 		if cmp.Equal(old, obj) {
 			if i == 0 {
 				return false, nil