Depends on #1452.

Add a basic benchmark test for System.Mutate.

In this case benchmarking pointed to an easy way to improve performance dramatically, so I've bundled the changes together in this PR.

Most of the time we spend in System.Mutate is in cmp.Equal. However, we know that "old" and "obj" are identical if no Mutators were run. This means if in an iteration no Mutators reported making any changes, we can immediately assume:

1) If we are on the first iteration, then old == obj == original, and so no changes were made to the object.
2) Otherwise, we know that (old == obj), and both are distinct from original (since we verified this in the first iteration).

```
$ benchstat before.txt after.txt 
name              old time/op  new time/op  delta
System_Mutate-12  8.77µs ± 3%  1.14µs ± 7%  -86.97%  (p=0.000 n=20+20)
```