@willbeason 2021-07-19
After this PR and #1439, it probably isn't worth doing further mutation-specific performance improvements. After these, almost 90% of our cpu is spent generating a UUID for the mutation, and only 10% is spent doing the actual mutation.